requirements.txt
torch>=2.0.0
torchvision
monai>=1.3.0
SimpleITK>=2.3.0
nibabel
numpy
scipy
scikit-image
matplotlib
tensorboard
pyyaml
tqdm
pillow
pydicom
# Note: pyradiomics has Python 3.13 compatibility issues
# Using custom radiomics extractor in src/utils/radiomics.py instead
